# Add index on clinic_activity_logs.numeric_value for performance optimization-created-by-agentic