--- a/src/main/resources/db/migration/V2_1__Add_Index_Clinic_Activity_Logs.sql
+++ b/src/main/resources/db/migration/V2_1__Add_Index_Clinic_Activity_Logs.sql
@@ -1,0 +1,2 @@
+-- Add index on numeric_value column to improve query performance
+CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs(numeric_value);