This PR addresses a critical performance issue in the Activity Logs Query endpoint.

Problem:
- Query performance degradation (5.18s vs expected 931μs)
- Missing index on numeric_value column causing sequential scans
- High business impact (criticality: 0.88)

Solution:
- Added index on clinic_activity_logs.numeric_value column
- Created database migration file to ensure index persistence
- Performance improvement verified (new execution time: ~0.2ms)

Related Issues:
- 5aa1fb7a-3656-11f0-8db3-ca59c7e8e81d
- 14dbd63a-4554-11f0-b280-42975a299f33

Verification:
- Index creation successful
- Query execution time improved from 5.18s to ~0.2ms
- No sequential scans after index implementation